Implement fake-able hrtime(), deprecate microtime()
ConvertibleTimestamp::microtime() was recently added and has exactly
one usage in Codesearch Everywhere (CheckUser extension), which should
use hrtime() instead.

I think it is generally true that if something relates to a real-world
date and time, seconds suffice. When something needs fractions, it is
either about entropy or about measuring a intra-process delta between
two points in time. Either way, you'd want hrtime instead.

Note that while hrtime() is safe to use (PHP 7.3+, we require PHP 7.4)
it is still relatively new which is why in most measuring code we
still use microtime today. But, we're trying to move away from this!

Bug: T245464
Bug: T340552
Change-Id: Ie48d9400cf3bb99467a16b9ea0cce6172e2e2568