Change getTimestamp() to throw InvalidArgumentException on bad format
Output formats are typically hardcoded, and at any rate, methods in this
library already require callers to use TS_* constants for them. Hence,
passing an invalid output format should never be tolerated, as it's a
direct violation of the method contract that the caller should be made
aware of. This scenario should not be treated the same as an invalid
timestamp (if user-provided, for example) or otherwise ignored as in
`::convert()`. As a case in point, if someone swaps the $style and $ts
parameters to convert(), they don't get a meaningful error message and
might assume that the timestamp was invalid.

So, replace (checked) TimestampException with (unchecked)
InvalidArgumentException, to ensure that the exception is never
suppressed.

Change-Id: I2eb32f9cb8561cc72622bdbc77f42954eced5503