Faster setTimestamp() by moving TS_UNIX up in the list of regexes
In a local test (with MySQL) the TS_MW format is the by far most
relevant one. It's already at the very top. Which means the loop
will usually stop after the very first iteration.

TS_UNIX appears to be the second most relevant. Moving it to the
second position means we can avoid a bunch of extra preg_match calls.
Since this loop is called a lot (sometimes a few hundred times per
request, e.g. on Special:RecentChanges) this adds up quickly.

The formats involved in this change are all exclusive to each other
and cannot overlap. In other words: I can guarantee that this change
doesn't make any difference, except performance.

Change-Id: If401dabe667245b1835e88e75cea5b141aa9fa5e